Fixes:
- Fixed gpt3 class and message max characters error
- Moved processors functions to processors.py
- Fixed kiwi analytics
- Minor cleanup of main.py
